[IDP-2311] add new test on TransactionInProgressRepository
